Write package config for vulkan_loader_and_layers package
This writes a GN package definition for the vulkan_loaders_and_layers
package and adds a new vulkan_loader_and_layers_tests package that
contains the test binaries when targeting x86-64. This also updates the
location of the icd.d / explicit_layer.d config files to
/system/data/vulkan instead of /system/lib/vulkan to be more consistent
with other data files.

Change-Id: Ic6e85773212ab68dae490cc4a23d6cbcc8eeacd9