Update loader fuchsia_shared_library rule for configs change
The fuchsia_shared_library template is being updated to treat configs
in the same way that the builtin shared_library does. This means the
configs variable is initialized when the rule evaluates, so
manipulations to the configurations have to manipulate the list insetad
of specify a new list that is appended.

Change-Id: I68010afcd7122858c205e3c1bfd8c9536f760bc7